in ExecuteALU.cpp, cast ops to 64bit before(!) shifting also fix more warnings found by static analysis and fix some 64bit compilation issues (much more to be done though) and properly return result for MediaGetInfo() also close files in MediaRead/Write() and LoadNuonRomFile() and InstallShaderSourceFromFile()
git-svn-id: https://svn.code.sf.net/p/nuance-nuon-emulator/code@54 6b0bf7a3-bffa-4ecf-8aa8-0eaf4365b716

+++ nativecodecache.h
@@ -53,8 +53,8 @@ public:
   {
     if(boundary)
     {
-      const uint32 poweroftwominusone = (1UL << boundary) - 1UL;
-      pEmitLoc = (uint8 *)(((uint32)(pEmitLoc + poweroftwominusone)) & ~(poweroftwominusone));
+      const size_t poweroftwominusone = (1UL << boundary) - 1UL;
+      pEmitLoc = (uint8 *)(((size_t)(pEmitLoc + poweroftwominusone)) & ~(poweroftwominusone));
     }
   }
